adding bootstrap and jquery and stylesheet

+++ WebApplication/BasicWebApplication/src/main/java/com/example/springboot/web/springbootfirstwebapp/service/TodoService.java
@@ -37,7 +37,7 @@ public class TodoService {
         todos.add(new Todo(++todoCount, name, desc, date, isDone));
     }
 
-    public void deleteTodo(int id){
+    public void deleteTodo(long id){
         Iterator<Todo> iterator = todos.iterator();
         while(iterator.hasNext()){
             Todo todo = iterator.next();
